refactor: split InsertBlockDialog
